Split Lens knowledge base UI
